+++ crates/cli/src/commands/mod.rs
@@ -1,5 +1,9 @@
 pub mod build;
-pub mod commands;
+pub mod clean;
+pub mod core;
 pub mod error;
+pub mod init;
 pub mod regtest;
 pub mod test;
+
+pub use core::*;