+++ crates/sdk/src/program/mod.rs
@@ -1,9 +1,9 @@
 pub mod arguments;
+pub mod core;
 pub mod error;
-pub mod program;
 pub mod witness;
 
 pub use arguments::ArgumentsTrait;
+pub use core::{Program, ProgramTrait};
 pub use error::ProgramError;
-pub use program::{Program, ProgramTrait};
 pub use witness::WitnessTrait;