+++ crates/sdk/src/signer/mod.rs
@@ -1,5 +1,5 @@
+pub mod core;
 pub mod error;
-pub mod signer;
 
+pub use core::{Signer, SignerTrait};
 pub use error::SignerError;
-pub use signer::{Signer, SignerTrait};