+++ crates/sdk/src/transaction/partial_input.rs
@@ -5,14 +5,14 @@ use simplicityhl::elements::{AssetId, OutPoint, Sequence, TxOut, Txid};
 use crate::program::ProgramTrait;
 use crate::program::WitnessTrait;
 
-#[derive(Clone)]
+#[derive(Debug, Clone)]
 pub enum RequiredSignature {
     None,
     NativeEcdsa,
     Witness(String),
 }
 
-#[derive(Clone)]
+#[derive(Debug, Clone)]
 pub struct PartialInput {
     pub witness_txid: Txid,
     pub witness_output_index: u32,
@@ -53,58 +53,53 @@ impl PartialInput {
             witness_txid: outpoint.txid,
             witness_output_index: outpoint.vout,
             witness_utxo: txout,
-            sequence: sequence,
-            amount: amount,
-            asset: asset,
+            sequence,
+            amount,
+            asset,
         }
     }
 
     pub fn outpoint(&self) -> OutPoint {
         OutPoint {
-            txid: self.witness_txid.clone(),
+            txid: self.witness_txid,
             vout: self.witness_output_index,
         }
     }
 
     pub fn input(&self) -> Input {
-        let mut input = Input::default();
-
-        input.previous_txid = self.witness_txid.clone();
-        input.previous_output_index = self.witness_output_index;
-        input.witness_utxo = Some(self.witness_utxo.clone());
-        input.sequence = Some(self.sequence.clone());
-        input.amount = self.amount.clone();
-        input.asset = self.asset.clone();
-
-        input
+        Input {
+            previous_txid: self.witness_txid,
+            previous_output_index: self.witness_output_index,
+            witness_utxo: Some(self.witness_utxo.clone()),
+            sequence: Some(self.sequence),
+            amount: self.amount,
+            asset: self.asset,
+            ..Default::default()
+        }
     }
 }
 
 impl ProgramInput {
     pub fn new(program: Box<dyn ProgramTrait>, witness: Box<dyn WitnessTrait>) -> Self {
-        Self {
-            program: program,
-            witness: witness,
-        }
+        Self { program, witness }
     }
 }
 
 impl IssuanceInput {
     pub fn new(issuance_amount: u64, asset_entropy: [u8; 32]) -> Self {
         Self {
-            issuance_amount: issuance_amount,
-            asset_entropy: asset_entropy,
+            issuance_amount,
+            asset_entropy,
         }
     }
 
     pub fn input(&self) -> Input {
-        let mut input = Input::default();
-
-        input.issuance_value_amount = Some(self.issuance_amount);
-        input.issuance_asset_entropy = Some(self.asset_entropy);
-        input.issuance_inflation_keys = None;
-        input.blinded_issuance = Some(0x00);
-
-        input
+        Input {
+            issuance_value_amount: Some(self.issuance_amount),
+            issuance_asset_entropy: Some(self.asset_entropy),
+            issuance_inflation_keys: None,
+            blinded_issuance: Some(0x00),
+            ..Default::default()
+        }
     }
 }